backend/requirements.txt
Flask
flask-cors
openai
python-dotenv
flask
flask_sqlalchemy
bcrypt
jwt
PyJWT
flask-cors
openai
python-dotenv
flask-migrate
flask-jwt-extended